flask-test/requirements.txt
Flask==2.0.2
azure-common==1.1.24
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.14.0
pylint==2.6.0
waitress==1.0.1
azure-cosmos==4.3.0
werkzeug